Removes redundent code, refactors LiftContainerApp and LiftLocalApp, and removes obselete containers
