+++ server/fig.yml
@@ -2,12 +2,12 @@ cassandra:
   image: spotify/cassandra
   expose:
     - "7199"
-    - "7000" 
-    - "7001" 
-    - "9160" 
-    - "9042" 
-    - "22" 
-    - "8012" 
+    - "7000"
+    - "7001"
+    - "9160"
+    - "9042"
+    - "22"
+    - "8012"
     - "61621"
 etcd:
   image: coreos/etcd
@@ -23,7 +23,6 @@ main:
     JOURNAL: journal.lift.eigengo.com
     SNAPSHOT: snapshot.lift.eigengo.com
     ETCD_URL: http://etcd.lift.eigengo.com:4001
-    KAFKA_BROKER_LIST: 192.168.59.103:9092
   ports:
     - "2552"
     - "8080"
@@ -33,54 +32,3 @@ main:
     - cassandra:journal.lift.eigengo.com
     - cassandra:snapshot.lift.eigengo.com
     - etcd:etcd.lift.eigengo.com
-    - zookeeper:zk
-adapter:
-  image: janm399/lift:adapter
-  environment:
-    ETCD_URL: http://etcd.lift.eigengo.com:4001
-  ports:
-    - "2552"
-  ports:
-    - "8080:8080"
-  volumes:
-    - /var/run/docker.sock:/var/run/docker.sock
-  links:
-    - cassandra:journal.lift.eigengo.com
-    - cassandra:snapshot.lift.eigengo.com
-    - etcd:etcd.lift.eigengo.com
-zookeeper:
-  image: wurstmeister/zookeeper
-  ports:
-    - "2181:2181"
-kafka:
-  image: wurstmeister/kafka:0.8.1.1-1
-  ports:
-    - "9092:9092"
-  links:
-    - zookeeper:zk
-  environment:
-    KAFKA_ADVERTISED_HOST_NAME: 192.168.59.103
-  volumes:
-    - /var/run/docker.sock:/var/run/docker.sock
-#sparkmaster:
-#  image: martinz/spark-master:latest
-#  ports:
-#    - "8080"
-#    - "7077:7077"
-#sparkworker:
-#  image: martinz/spark-worker:latest
-#  ports:
-#    - "8081:8081"
-#    - "8888:8888"
-#  links:
-#    - sparkmaster
-#sparkdriver:
-#  image: janm399/lift:spark
-#  ports:
-#    - "4040:4040"
-#    - "9001:9001"
-#  links:
-#    - sparkmaster
-#    - zookeeper:zk
-#    - cassandra:journal.lift.eigengo.com
-#    - cassandra:snapshot.lift.eigengo.com
